Create Test_UI_1
I'm trying to display user input data from one page on a separate page for future calculation purposes.  However, user input data entered on one page is not saved and transferred onto the next page.  I know this is because my pushbutton runs the windowOne function and returns data from an empty textbox, but what I don't know is how to make that not happen.